Update README.md
I'm a doctor and I had trouble understanding the goal of the project. Therefore, I made some changes to increase the readability, make the file more understandable, and strengthen the idea behind the project. I also should add that I would definitely want to talk about and work on the project, if possible, since it is one of the most creative ML projects in medicine I have seen so far.